Resolved not to delete the default image

+++ BlogApp/repository/database_image_repo.py
@@ -3,7 +3,7 @@ import uuid
 from repository.image_repo import ImageRepo
 
 IMG_FOLDER_PATH = 'static/img/'
-
+DEFAULT_IMG = '1.jpg'
 class DatabaseImageRepo(ImageRepo):
     def add(self, file_storage):
         if file_storage.filename != '':
@@ -20,7 +20,7 @@ class DatabaseImageRepo(ImageRepo):
         return self.add(file_storage)
 
     def delete(self, filename):
-        if os.path.isfile(IMG_FOLDER_PATH + '{}'.format(filename)):
+        if filename != DEFAULT_IMG and os.path.isfile(IMG_FOLDER_PATH + '{}'.format(filename)):
             os.remove(IMG_FOLDER_PATH + '{}'.format(filename))
         else:
             print("Error: %s file not found" % filename)
